# services/rest/api/openapi.yml
openapi: 3.1.0
servers:
  - url: 'https://localhost:8010/api/rest'
    description: Local LunaOne full node's API sever with TLS
  - url: 'https://testnet1.lunaone.com'
    description: Testnet 1 API sever to test features waiting for release
info:
  description: >-
    Provides REST API endpoints to work with LunaOne node by sending/viewing
    transactions, accounts, contracts and node state
  version: 1.0.0
  title: LunaOne full node API
  termsOfService: 'https://to.be.defined'
  contact:
    email: support@aplfintech.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /transactions/{txHash}:
    get:
      security: 
        - basic: []
      summary: Get confirmed transaction by its hash
      description: >-
        Fetch the confirmed transaction by its hexadecimal hash. Unconfirmed
        transaction will not be returned here
      parameters:
        - name: txHash
          in: path
          required: true
          schema:
            type: string
          description: >-
            Hash of the confirmed transaction to fetch in hexadecimal format
            prefixed with '0x'
      operationId: getTransactionByHash
      responses:
        '200':
          description: Found confirmed transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedTransaction'
        '404':
          description: Confirmed tx not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - transactions
  /transactions:
    post:
      security: 
        - basic: []
      summary: Send signed transaction
      description: >-
        Send raw signed transaction encoded in hexadecimal format to be accepted
        in mempool of the full node and then distributed withing the LunaOne
        network and confirmed by validators
      parameters: []
      operationId: sendRawTransaction
      responses:
        '200':
          description: >-
            Transaction successfully parsed, validated and accepted into a
            mempool
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TxReceipt'
        '400':
          description: >-
            Transaction validation error: bad signature, transaction cannot be
            parsed, transaction version is outdated, incorrect chain id
            specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: >-
            Needed entity for sending tx doesn't exist: sender's account,
            contract to be called
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: >-
            Conflict between node's state and client's state, required state
            repeated fetch by client or additional synchronization: bad nonce
            for transaction, not enough balance on account to pay fees+amount 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RawTransaction'
        description: >-
          Signed transaction encoded in hexadecimal format and prefixed with
          '0x'. Tx structure and encoding rules are defined in a separate
          documents

    get:
      security: 
        - basic: []
      summary: Get confirmed and unconfirmed transactions with filter
      description: >-
        Return list of confirmed or/and unconfirmed transactions using pagination and optinal filter by account, 
        sender/recipient with an optional sort by timestamp/nonce DESC/ASC
      parameters:
        - in: query
          name: accounts
          schema:
            type: array
            items:
              type: string
              format: address
          style: form
          explode: false
          example: ["0xd09ec4a81cde61b57de012d3fe80beae3f28fb68"]
          description: List of accounts that must be used for filtering transactions, when empty - then all transactions will be returned
        - in: query
          name: block
          description: Number of the block to see for transactions (optional)
          schema: 
            type: integer
            minimum: 0
          examples: 
            genesis:
              value: 0
              summary: Transactions from the genesis block (typically genesis has no transactions)
            arbitrary_block:
              value: 1250
              summary: Get transaction from block with number 1250 (1250th block after genesis)
        - in: query
          name: direction
          schema:
            type: string
            enum: 
            - SENT
            - RECEIVED
            - BOTH
            default: BOTH
          examples: 
            all:
              value: BOTH
              summary: Don't filter out transaction by SENT/RECEIVED criteria, keep both types, default
            sent:
              value: SENT
              summary: Return only SENT transactions, skipping all received
            received: 
              value: RECEIVED
              summary: Return only RECEIVED transactons, skipping all sent
        - in: query
          name: sortOrder
          schema: 
            type: string
            enum: 
            - ASC
            - DESC
            default: DESC
          description: Sort fetched transactions in ascending or descending order by 'sortCriteria'
          example: ASC
        - in: query
          name: sortCriteria
          schema: 
            type: string
            enum: 
            - TIMESTAMP
            - NONCE
            default: NONCE
          examples:
            nonceSort:
              value: NONCE
              summary: Sort transactions by their order - nonce
            timestampSort: 
              value: TIMESTAMP
              summary: Sort transactions by the time of their creation historical order
          description: Criteria to sort transactions by, NONCE is suited for one account transactions ordering, TIMESTAMP for many accounts ordering
        - in: query
          name: txType
          schema:
            type: integer
            enum:
              - -1
              - 0
              - 1
              - 2
            default: -1
          description: Filter transactions by type, leave empty or -1 to not apply this filter
          examples:
              paymentType:
                value: 0
                summary: Fetch only payment txs
              publishType:
                value: 1
                summary: Fetch only PublishContract txs
              callType:
                value: 2
                summary: Fetch only CallContract txs
        - in: query
          name: confirmationStatus
          description: Select transactions with statuses - Confirmed, Unconfirmed or Both
          schema:
            type: string
            enum: 
            - CONFIRMED
            - UNCONFIRMED
            - BOTH
            default: BOTH
          examples: 
            all:
              value: BOTH
              summary: Select both unconfirmed and confirmed transactions
            sent:
              value: UNCONFIRMED
              summary: Select only unconfirmed transactions
            received: 
              value: CONFIRMED
              summary: Select only confirmed transactions
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/pageSizeParam'
      operationId: getFilteredTransactions
      responses:
        '200':
          description: Filtered, sorted, paginated transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedTransactions'
        '400':
          description: Incorrect sorting/filtering/paginating parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - transactions
  /transactions/estimate:      
    post:
      security: 
        - basic: []
      summary: Estimate fuel limit for tx
      description: >-
        Estimate fuel limit for transaction execution by executing this mocked tx against current
        state on VM of the p2p node and then reverting back to initial state (no state changes)
        Two inputs are possible: RawTransaction (unsigned), CallMessage (sender, recipient, amount, data)
      parameters: []
      operationId: estimateTxFuel
      responses:
        '200':
          description: >-
            Transaction estimation result (via executing mocked tx). Can display validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuelEstimationResult'
        '400':
          description: >-
            Transaction cannot be parsed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CallMessage'
                - $ref: '#/components/schemas/RawTransaction'
        description: >-
          Unsigned transaction encoded in hexadecimal format and prefixed with
          '0x'. Tx structure and encoding rules are defined in a separate
          documents
...
components:
  parameters: 
    pageParam:
      in: query
      name: page
      schema: 
        type: integer
        minimum: 0
        default: 0
      description: Number of the page to be returned, default is 0 (first page)
      examples: 
        first_page:
          value: 0
          summary: "First page number, when is not set, then first page will be used by default"
        10th_page:
          value: 9
          summary: "10th page, starting from 9 * pageSize element up to 10 * pageSize (exlusive)"
    pageSizeParam:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 15
      description: Size of the page returned. Typically mustn't be changed on paging through because it will be required to recalculate position of the first element = page * size
      examples: 
        min_size:
          value: 1
          summary: The smallest page to return consist of the only one element
        typical_size:
          value: 25
          summary: Return page containing 25 elements
  schemas:
...
    UnifiedTransactions:
      type: object
      description: List of transactions sorted by certain criteria
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/UnifiedTransaction'
...
    ApiError:
      type: object
      description: General API error returned when any error occurred during performing API call
      properties:
        error:
          type: string
          description: Reason of the occurred error or its content
          example: 'API error occurred, reason: Bad request'
    RawTransaction:
      type: object
      description: New signed transaction to be sent to the LunaOne network
      properties:
        encodedTx: 
          type: string
          format: hexadecimal
          description: Signed and encoded into a the hexadecimal format transaction that must be sent into the LunaOne network and get confirmed
          example: "0x12321fcbdaaadd12977a7a802000002371232"
      required:
        - encodedTx
    TxReceipt:
      type: object
      description: Transaction execution receipt object
      properties:
        txHash: 
          type: string
          format: hexadecimal
          minLength: 66
          maxLength: 66
          description: Hash of the transaction (Keccak-256) prefixed with '0x'
          example: "0xdf12984738297358323203203743643242389dacbdf23"
        executionStatus:
          type: string
          description: Status of transaction execution denoting either successful execution ('SUCCESS') or failed (with possible error cause) - any other output
          example: 'Nonce is too low'
        fuelUsed:
          type: integer
          format: int64
          description: Amount of fuel used for transaction execution out of specified fuel limit
          example: 120545    
...
  securitySchemes:
    adminAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: ADMIN authrorization scheme to get access to system admin endpoints
    basic:
      type: http
      scheme: basic
...
security: []